requirements.txt
Django>=1.11,<3.0
djangorestframework>=3.6.4
